backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
geopy==2.4.1
scikit-learn==1.4.2
numpy==1.26.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0
